Refactor Axvisor and Starry test implementations
- Moved QEMU test logic from Axvisor to Starry, consolidating test functions.
- Removed redundant code and improved structure for better readability and maintainability.
- Updated test handling for U-Boot and board tests in both Axvisor and Starry.
- Enhanced error handling and logging for test execution.
- Streamlined configuration loading and request preparation for tests.
