+++ scripts/axbuild/src/axvisor/test.rs
@@ -1,14 +1,21 @@
 use std::{
     fs,
     path::{Path, PathBuf},
+    time::Instant,
 };
 
 use anyhow::{Context, anyhow, bail};
+use ostool::{board::RunBoardOptions, build::config::Cargo, run::qemu::QemuConfig};
 use serde::Deserialize;
 
+use super::{Axvisor, build, cli, rootfs};
 use crate::{
-    context::resolve_axvisor_arch_and_target,
-    test::{case::TestQemuCase, qemu::parse_test_target},
+    command_flow::SnapshotPersistence,
+    context::{AxvisorCliArgs, ResolvedAxvisorRequest, resolve_axvisor_arch_and_target},
+    test::{
+        board as board_test, case as test_case, case::TestQemuCase, qemu as test_qemu,
+        qemu::parse_test_target,
+    },
 };
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -440,6 +447,309 @@ fn supported_test_groups(test_suite_root: &Path) -> anyhow::Result<String> {
 
 fn qemu_config_name(arch: &str) -> String {
     format!("qemu-{arch}.toml")
+}
+
+impl Axvisor {
+    pub(super) async fn test_qemu(&mut self, args: cli::ArgsTestQemu) -> anyhow::Result<()> {
+        let (arch, target) = parse_target(&args.arch, &args.target)?;
+        let cases = discover_qemu_cases(
+            self.app.workspace_root(),
+            &args.test_group,
+            &arch,
+            args.test_case.as_deref(),
+        )?;
+
+        println!(
+            "running axvisor qemu tests for arch: {} (target: {}, cases: {})",
+            arch,
+            target,
+            cases.len()
+        );
+
+        let build_config = qemu_test_build_config(&cases)?;
+        let vmconfigs = qemu_test_vmconfigs(&cases);
+
+        let request = self.prepare_request(
+            axvisor_qemu_test_build_args(&arch, build_config, vmconfigs),
+            None,
+            None,
+            SnapshotPersistence::Discard,
+        )?;
+        rootfs::ensure_qemu_rootfs_ready(&request, self.app.workspace_root(), None).await?;
+        let cargo = build::load_cargo_config(&request)?;
+        self.app.set_debug_mode(request.debug)?;
+        self.app
+            .build(cargo.clone(), request.build_info_path.clone())
+            .await
+            .context("failed to build shared Axvisor qemu test artifact")?;
+
+        let total = cases.len();
+        let suite_started = Instant::now();
+        let mut failed = Vec::new();
+        for (index, case) in cases.iter().enumerate() {
+            println!("[{}/{}] axvisor qemu {}", index + 1, total, case.case.name);
+
+            let case_started = Instant::now();
+            let result = self
+                .run_qemu_case(&request, &cargo, case)
+                .await
+                .with_context(|| format!("axvisor qemu test failed for case `{}`", case.case.name));
+            match result {
+                Ok(()) => println!("ok: {} ({:.2?})", case.case.name, case_started.elapsed()),
+                Err(err) => {
+                    eprintln!("failed: {}: {err:#}", case.case.name);
+                    failed.push(case.case.name.clone());
+                }
+            }
+        }
+
+        println!("axvisor qemu total: {:.2?}", suite_started.elapsed());
+        test_qemu::finalize_qemu_test_run("axvisor", &failed)
+    }
+
+    pub(super) async fn test_uboot(&mut self, args: cli::ArgsTestUboot) -> anyhow::Result<()> {
+        let board = uboot_board_config(&args.board, &args.guest)?;
+        let explicit_uboot_config = args.uboot_config.clone();
+        let uboot_config_summary = explicit_uboot_config
+            .as_ref()
+            .map(|path| path.display().to_string())
+            .unwrap_or_else(|| "using ostool default search".to_string());
+
+        if let Some(path) = explicit_uboot_config.as_ref()
+            && !path.exists()
+        {
+            bail!(
+                "missing explicit U-Boot config `{}` for axvisor board tests",
+                path.display()
+            );
+        }
+
+        println!(
+            "running axvisor uboot test for board: {} guest: {} with vmconfig: {}",
+            board.board, board.guest, board.vmconfig
+        );
+
+        let mut request = self.prepare_request(
+            axvisor_uboot_test_build_args(board.build_config, board.vmconfig),
+            None,
+            explicit_uboot_config.clone(),
+            SnapshotPersistence::Discard,
+        )?;
+        request.uboot_config = explicit_uboot_config;
+
+        let cargo = build::load_cargo_config(&request)?;
+        let uboot = self.load_uboot_config(&request, &cargo).await?;
+        self.app
+            .uboot(cargo, request.build_info_path, uboot)
+            .await
+            .with_context(|| {
+                format!(
+                    "axvisor uboot test failed for board `{}` guest `{}` (build_config={}, \
+                     vmconfig={}, uboot_config={})",
+                    board.board,
+                    board.guest,
+                    board.build_config,
+                    board.vmconfig,
+                    uboot_config_summary
+                )
+            })
+    }
+
+    pub(super) async fn test_board(&mut self, args: cli::ArgsTestBoard) -> anyhow::Result<()> {
+        let groups = discover_board_test_groups(
+            self.app.workspace_root(),
+            &args.test_group,
+            args.test_case.as_deref(),
+            args.board.as_deref(),
+        )?;
+        let total = groups.len();
+        let mut failed = Vec::new();
+
+        for (index, group) in groups.into_iter().enumerate() {
+            let group_label = format!("{}/{}", group.name, group.board_name);
+            let board_test_config = group.board_test_config_path.clone();
+            let board_test_config_summary = board_test_config.display().to_string();
+
+            if !board_test_config.exists() {
+                eprintln!(
+                    "failed: {}: missing board test config `{}`",
+                    group_label, board_test_config_summary
+                );
+                failed.push(group_label);
+                continue;
+            }
+
+            println!("[{}/{}] axvisor board {}", index + 1, total, group_label);
+
+            let result = async {
+                let prepared_vmconfigs = group.vmconfigs.clone();
+                let request = self.prepare_request(
+                    axvisor_board_test_build_args(&group, prepared_vmconfigs),
+                    None,
+                    None,
+                    SnapshotPersistence::Discard,
+                )?;
+                let cargo = build::load_cargo_config(&request)?;
+                let board_config = self
+                    .load_board_config(&cargo, Some(board_test_config.as_path()))
+                    .await?;
+                self.app
+                    .board(
+                        cargo,
+                        request.build_info_path,
+                        board_config,
+                        RunBoardOptions {
+                            board_type: args.board_type.clone(),
+                            server: args.server.clone(),
+                            port: args.port,
+                        },
+                    )
+                    .await
+                    .with_context(|| {
+                        format!(
+                            "axvisor board test failed for group `{}` (build_config={}, \
+                             board_test_config={}, vmconfigs={})",
+                            group_label,
+                            group.build_config.display(),
+                            board_test_config_summary,
+                            group
+                                .vmconfigs
+                                .iter()
+                                .map(|path| path.display().to_string())
+                                .collect::<Vec<_>>()
+                                .join(", ")
+                        )
+                    })
+            }
+            .await;
+
+            match result {
+                Ok(()) => println!("ok: {}", group_label),
+                Err(err) => {
+                    eprintln!("failed: {}: {:#}", group_label, err);
+                    failed.push(group_label);
+                }
+            }
+        }
+
+        board_test::finalize_board_test_run("axvisor", &failed)
+    }
+
+    async fn load_qemu_case_config(
+        &mut self,
+        request: &ResolvedAxvisorRequest,
+        cargo: &Cargo,
+        case: &AxvisorQemuCase,
+    ) -> anyhow::Result<QemuConfig> {
+        let mut qemu = self
+            .app
+            .tool_mut()
+            .read_qemu_config_from_path_for_cargo(cargo, &case.case.qemu_config_path)
+            .await?;
+        test_case::apply_grouped_qemu_config(&mut qemu, &case.case);
+
+        let mut case_request = request.clone();
+        case_request.vmconfigs = case.vmconfigs.clone();
+        let rootfs_path = rootfs::qemu_rootfs_path(&case_request, self.app.workspace_root(), None)?;
+        let prepared_assets = test_case::prepare_case_assets(
+            self.app.workspace_root(),
+            &case_request.arch,
+            &case_request.target,
+            &case.case,
+            rootfs_path,
+        )
+        .await?;
+        rootfs::patch_qemu_rootfs_path(&mut qemu, &prepared_assets.rootfs_path);
+        qemu.args.extend(prepared_assets.extra_qemu_args);
+        Ok(qemu)
+    }
+
+    async fn run_qemu_case(
+        &mut self,
+        request: &ResolvedAxvisorRequest,
+        cargo: &Cargo,
+        case: &AxvisorQemuCase,
+    ) -> anyhow::Result<()> {
+        let qemu = self.load_qemu_case_config(request, cargo, case).await?;
+        self.app.run_qemu(cargo, qemu).await
+    }
+}
+
+fn axvisor_qemu_test_build_args(
+    arch: &str,
+    config: Option<PathBuf>,
+    vmconfigs: Vec<PathBuf>,
+) -> AxvisorCliArgs {
+    AxvisorCliArgs {
+        config,
+        arch: Some(arch.to_string()),
+        target: None,
+        plat_dyn: None,
+        smp: None,
+        debug: false,
+        vmconfigs,
+    }
+}
+
+fn qemu_test_build_config(cases: &[AxvisorQemuCase]) -> anyhow::Result<Option<PathBuf>> {
+    let mut build_config: Option<PathBuf> = None;
+    for case in cases {
+        let Some(next) = &case.build_config else {
+            continue;
+        };
+        if let Some(current) = &build_config
+            && current != next
+        {
+            anyhow::bail!(
+                "Axvisor qemu cases in one run must use the same build_config for build-once; \
+                 `{}` uses `{}`, but an earlier case uses `{}`",
+                case.case.name,
+                next.display(),
+                current.display()
+            );
+        }
+        build_config = Some(next.clone());
+    }
+    Ok(build_config)
+}
+
+fn qemu_test_vmconfigs(cases: &[AxvisorQemuCase]) -> Vec<PathBuf> {
+    let mut vmconfigs = Vec::new();
+    for case in cases {
+        for vmconfig in &case.vmconfigs {
+            if !vmconfigs.contains(vmconfig) {
+                vmconfigs.push(vmconfig.clone());
+            }
+        }
+    }
+    vmconfigs
+}
+
+fn axvisor_uboot_test_build_args(build_config: &str, vmconfig: &str) -> AxvisorCliArgs {
+    AxvisorCliArgs {
+        config: Some(PathBuf::from(build_config)),
+        arch: None,
+        target: None,
+        plat_dyn: None,
+        smp: None,
+        debug: false,
+        vmconfigs: vec![PathBuf::from(vmconfig)],
+    }
+}
+
+fn axvisor_board_test_build_args(
+    group: &BoardTestGroup,
+    vmconfigs: Vec<PathBuf>,
+) -> AxvisorCliArgs {
+    AxvisorCliArgs {
+        config: Some(group.build_config.clone()),
+        arch: None,
+        target: None,
+        plat_dyn: None,
+        smp: None,
+        debug: false,
+        vmconfigs,
+    }
 }
 
 #[cfg(test)]
